requirements.txt
Django==2.1.5
djangorestframework==3.9.1
gunicorn==19.9.0
pytz==2018.9
six==1.12.0
psycopg2==2.7.7
celery==4.2.1
requests==2.21.0
